Added structure to ast mod
