+++ src/ast/m_integer.rs
@@ -1,0 +1,17 @@
+use crate::ast::{OperatorApplyMatcher, Matcher, Value, EvalError};
+
+impl OperatorApplyMatcher for Matcher<'_, i32> {
+    fn name(&self) -> &'static str {
+        "Integer"
+    }
+
+    fn apply_plus_with_integer(&self, other: &i32) -> Result<Value, EvalError> {
+        Ok(Value::Integer(self.value + other))
+    }
+    fn apply_plus_with_string(&self, other: &String) -> Result<Value, EvalError> {
+        let mut new = String::new();
+        new.push_str(self.value.to_string().as_str());
+        new.push_str(other.as_str());
+        Ok(Value::String(new))
+    }
+}
